refactor(issue): standardize DTO structure and domain types (Phase 1f)
- Move IssueFiles from issueService to types/issue/issue-files.ts
- Create CreateIssueRequest and UpdateIssueRequest DTO stubs
- Centralize issue type imports through the domain layer index
- Update issue service and mutation hooks to import from domain index
