+++ types/issue/index.ts
@@ -1,4 +1,7 @@
 export * from './issue-comment';
+export * from './issue-files';
 export * from './issue-status';
 export * from './issue-type';
 export * from './issue';
+export * from './issue-create';
+export * from './issue-update';
