+++ types/issue/issue-files.ts
@@ -1,0 +1,3 @@
+export interface IssueFiles {
+  attachments?: File[];
+}
